update 6D resume plot
Most likely, the user will stop the 6D motion between two values (x0 and x1) and systematically, position x1 is sent to the controller. 'monitorData' function considers that the position of the robot is at the first position of the new list but the actual position is between x0 and x1. The travel time causes a latency while drawing the plots.